Deviation of Input Data from Euler Pole Solution

Calculates angular distance of the input points to the small or great circle of the Euler pole solution

Usage

euler_deviation(x, ep, sm = TRUE)

Arguments

x

either a matrix or data.frame containing the geographical coordinates, or a sf object.

ep

three column vector. Euler pole solution

sm

logical. Whether the solution is for a small circle or great circle geometries.

Value

numeric (angle in degrees). Angular difference of data points to the small or great circle.

Examples

data(tintina)
res1 <- euler_reco(tintina)
euler_deviation(tintina, res1)
#>  [1]  0.496405362  0.605133187  0.693854551  0.737723380  0.723030390
#>  [6]  0.704650018  0.662972237  0.491606121  0.437316075  0.421862797
#> [11]  0.390430783  0.358409151  0.495925555  0.362508124  0.263592215
#> [16]  0.146776241  0.086000870 -0.072237351 -0.108031756 -0.179054424
#> [21] -0.169412665 -0.298109247 -0.595997279 -0.963922843 -1.048347651
#> [26] -1.111203709 -1.138149360 -1.112120562  0.569581949  0.261327105
#> [31]  0.322477458  0.162888184 -0.092539972 -0.204082124 -0.547749422
#> [36] -0.579556234 -0.555452691 -0.536205084 -0.455176780 -0.456026400
#> [41] -0.392161774 -0.376711600 -0.309269856 -0.253069456 -0.108795921
#> [46] -0.084591199 -0.002049965 -0.018899895 -0.027727967 -0.025067534
#> [51] -0.022323854 -0.018840586 -0.004289172 -0.010429215 -0.023355380
#> [56] -0.039515097 -0.046770545 -0.040274204 -0.054363245 -0.011308925
#> [61]  0.153767873  0.282488525  0.333392145  0.356611565  0.373259185
#> [66]  0.396062428  0.452937764  0.495703340

data(south_atlantic)
res2 <- euler_reco(south_atlantic)
euler_deviation(south_atlantic, res2)
#>  [1] -0.1566025298  0.1357934338  0.1421647535  0.0512709308  0.0191809264
#>  [6]  0.0094475915 -0.0008121112 -0.0864417126 -0.1985551298 -0.1508811708
#> [11]  0.0320651336  0.2042653645