Best Euler Pole Solution for Shmax Data

Iteratively refines the Euler pole solution for a given set of \(\sigma_{Hmax}\) data by densifying the grid and selecting the Euler pole with the lowest dispersion.

Usage

euler_from_stress(x, iter = 3, fact = 2, ...)

Arguments

x

data.frame containing the coordinates of the point(s) (lat, lon), the direction of \(\sigma_{Hmax}\) azi and its standard deviation unc (optional)

iter

positive integer. Number of iterations to refine the Euler pole solution

fact

positive integer. Aggregation factor to increase number of cells in each direction (horizontally and vertically, applied in each iteration.

...

optional arguments passed to euler_stress_dispersion()

Value

sf POINT object

Examples

if (FALSE) { # \dontrun{
data("san_andreas", package = "tectonicr")
euler_from_stress(san_andreas, prd = 135)
} # }