Euler pole solution for geological structures

Finds the best Euler pole solution for a set of geographic locations that are aligned along a geological structure.

Usage

euler_reco(x, sc = TRUE, densify.x = FALSE, ...)

Arguments

x

either a matrix or data.frame containing the geographical coordinates, or a sf object.

sc

logical. Whether the structure described by the points x is expected to follow small (TRUE) or great circle (FALSE) arcs?

densify.x

logical. Whether the points x along the lines structure should be densified before analysis (TRUE) or not (FALSE, the default). Densification would yield equally spaced points along the line.

...

optional arguments passed to smoothr::densify() (only if densify.x = TRUE).

Value

numeric vector given the latitude (in degrees), longitude (in degrees), the misfit (0 - low, 1 - high), and (for small circle structures) the apical angle (in degrees) of the best fit Euler pole.

Details

Based on Gray, N.H., Geiser, P.A., Geiser, J.R. (1980). On the least-square fit of small and great circles to spherically projected data. Mathematical Geology, Vol. 12, No. 3, 1980

Examples

test <- smallcircle(30, 150, 35)
euler_reco(test)
#>          lat          lon        angle       misfit 
#> 3.000000e+01 1.500000e+02 3.500000e+01 5.780875e-11 

# Example from Price and Carmicheal (1986), GEOLOGY:

# from matrix
rmt_mat <- rbind(
  "yukon" = c(66.1, -147.8),
  "bigbend" = c(52.25, -122.65),
  "washington" = c(47.85, -121.85)
)
euler_reco(rmt_mat)
#>           lat           lon         angle        misfit 
#> -4.169012e+01  2.392828e+01  2.484264e+01  2.879342e-29 

# from data.frame
rmt_df <- data.frame("lat" = rmt_mat[, 1], "lon" = rmt_mat[, 2])
euler_reco(rmt_df)
#>           lat           lon         angle        misfit 
#> -4.169012e+01  2.392828e+01  2.484264e+01  2.879342e-29 

# from sf object
rmt_sf <- sf::st_as_sf(rmt_df, coords = c("lon", "lat"), crs = "WGS84")
euler_reco(rmt_sf)
#>           lat           lon         angle        misfit 
#> -4.169012e+01  2.392828e+01  2.484264e+01  2.879342e-29 

data(tintina)
euler_reco(tintina)
#>           lat           lon         angle        misfit 
#>  2.508365e+01 -1.598391e+02  4.111992e+01  4.440427e-03 
euler_reco(tintina, densify.x = TRUE)
#>           lat           lon         angle        misfit 
#>   24.65028141 -160.15652881   41.62510809    0.04028761 

data(south_atlantic)
euler_reco(south_atlantic, sc = TRUE)
#>           lat           lon         angle        misfit 
#> -6.040911e+01  1.449658e+02  6.012862e+01  5.442767e-05 
euler_reco(south_atlantic, densify.x = TRUE, sc = TRUE)
#>           lat           lon         angle        misfit 
#> -5.973840e+01  1.456660e+02  5.940073e+01  3.309771e-04